worker/requirements.txt
redis>=5.0.4,<6.0.0
rq>=1.16.2,<2.0.0
opentelemetry-api>=1.39.1,<2.0.0
opentelemetry-sdk>=1.39.1,<2.0.0
opentelemetry-exporter-otlp>=1.39.1,<2.0.0
opentelemetry-instrumentation-redis>=0.60b1,<1.0.0
